requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dateparser==1.1.1
fastapi==0.75.2
flake8==4.0.1
frozenlist==1.3.0
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
orjson==3.6.8
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.17.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
websockets==10.3
yarl==1.7.2
